Modularized the integration tests
Created a class IntegrationRest which contains a different database than the one in production and a httpClient
Created StudentsControllerTests and CoursesControllerTests which inherit the IntegrationTest class
Added tests for each Get methods
